+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update root namespace to `FiveOrbs`
+- Update root namespace to `Duon`
 
 ## [1.1.4] - 2024-08-24
 
@@ -75,4 +75,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- php-cs-fixer rules via `FiveOrbs\Development\PhpCsFixer\Config`.
+- php-cs-fixer rules via `Duon\Development\PhpCsFixer\Config`.